+++ src/opnsense/mvc/app/controllers/OPNsense/Abuseipdb/IndexController.php
@@ -14,6 +14,7 @@ class IndexController extends \OPNsense\Base\IndexController
         $this->view->generalForm = $this->getForm("general");
         $this->view->blacklistForm = $this->getForm("blacklist");
         $this->view->reporterForm = $this->getForm("reporter");
+        $this->view->suricataForm = $this->getForm("suricata");
         $this->view->selfcareForm = $this->getForm("selfcare");
         $this->view->permabanForm = $this->getForm("permaban");
         $this->view->pick('OPNsense/Abuseipdb/index');